Use postgres hstore for changes
(via activerecord-postgres-hstore)

This is totally the wrong way to go about it, and I'm a little shocked the internet hasn't figured out the right way, but my googling fails to find an example.

Really this should be configurable somehow, and the gem should gracefully depend on activerecord-postgres-hstore if available, and still function if not.

To use this as is, you must include activerecord-postgres-hstore in your Gemfile *before* the inclusion of audited-activerecord.

Grim, but don't tell me I didn't warn you.